refactor: extract NormalizeOptions into dedicated `options` module
Moves `NormalizeOptions`, its builder methods, and `DEFAULT_MAX_SPAN_TOKENS`
out of `src/lib.rs` and into a new `src/options.rs`. The struct is the
extension point for caller-tunable normalization behavior, and giving it
its own module makes room for richer per-field documentation and future
flags without further bloating the crate root.

Each field now carries:
- a "Default" line stating the no-op behavior
- a bulleted list of concrete input → output examples
- the originating issue number for the behavior
- guidance on which use cases want it on/off
- explicit interaction notes (which other taggers still win)

Also documents the `with_*` builder convention as the preferred construction
path so new fields can land without breaking existing call sites that use
struct literals.

`pub use options::{NormalizeOptions, DEFAULT_MAX_SPAN_TOKENS};` keeps the
public path stable — no changes required in FFI, WASM, Swift, or test code.
All Rust + FFI tests pass; WASM check clean.
